Refactor tsweb/src/k0rdent_catalog.tsx
